First pass at a some sorta scheduler
Not particularly useful right now. Most helpful thing it does is
calculate the the start dates of each task. Next thing to do is to
calculate an "id" of sorts, which indicates who the parent task is;
something like "2Ai" meaning task i of parent task 2A (which in turn
means task A of parent task 2).

A possible improvement may be to have ACTUAL scheduling of tasks,
where the duration of a task is decided by whether it is marked as
:easy or :hard, etc.; say e.g. a :hard task may require 3 times to
time that an :easy task would need. Ultimately I don't think this will
be a useful change: at some point will need to have finer control of
task duration.

It definitely needs cleaning up, though. It's "short" but it ain't
pretty. Unlike me -- I'm short AND pretty.